Split out integration tests
