--- .github/workflows/run-frontend-tests.yml
@@ -1,42 +1,0 @@
-name: Frontend unit tests
-
-on:
-  push:
-    branches:
-      - v1
-      - main
-  pull_request:
-    branches:
-      - v1
-      - main
-  schedule:
-    - cron: '0 9 1 * *'
-
-jobs:
-  run:
-    runs-on: ubuntu-latest
-    strategy:
-      fail-fast: true
-
-    steps:
-      - uses: actions/checkout@v5
-        with:
-          persist-credentials: false
-
-      - name: Setup Node.js environment
-        uses: actions/setup-node@v5
-        with:
-          node-version-file: '.nvmrc'
-
-      - name: Install yarn dependencies
-        run: yarn install
-        env:
-          NODE_OPTIONS: '--max_old_space_size=4096'
-
-      - name: Check types
-        run: yarn typecheck
-
-      - name: Build Frontend
-        run: yarn build
-        env:
-          NODE_OPTIONS: '--max_old_space_size=4096'
