--- config/custom.21.8.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" ?>
-<yandex>
-    <listen_host>::</listen_host>
-    <listen_host>0.0.0.0</listen_host>
-    <listen_try>1</listen_try>
-    <logger>
-        <console>1</console>
-    </logger>
-</yandex>
